refactor campaign approval inbox router
