+++ src/app/router_registry.py
@@ -171,6 +171,9 @@ from app.routers.dpm_wave_ai import router as dpm_wave_ai_router
 from app.routers.dpm_wave_campaign_approval_commands import (
     router as dpm_wave_campaign_approval_commands_router,
 )
+from app.routers.dpm_wave_campaign_approval_inbox import (
+    router as dpm_wave_campaign_approval_inbox_router,
+)
 from app.routers.dpm_wave_campaign_approvals import (
     router as dpm_wave_campaign_approvals_router,
 )
@@ -470,6 +473,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_wave_campaign_discovery_router)
     app.include_router(dpm_wave_campaign_workflow_boards_router)
     app.include_router(dpm_wave_campaign_assignment_views_router)
+    app.include_router(dpm_wave_campaign_approval_inbox_router)
     app.include_router(dpm_wave_campaign_approvals_router)
     app.include_router(dpm_wave_campaign_approval_commands_router)
     app.include_router(dpm_wave_campaign_assignments_router)
